cleanup (are we cooked)

+++ src/main/java/frc/robot/Subsystems/Shooter/ShooterConstants.java
@@ -24,7 +24,7 @@ public class ShooterConstants {
 
   public static final int currentLimit = 40;
 
-  public static final int indexerVoltage = 7;
+  public static final double indexerVoltage = 7.0;
 
   public static final double tolerableRpmRangeShooter = 10;
